cilium, recorder: add middle layer implementation
Add the core part of the recorder handling, that is, tracking of
recorder entries as well as mask generation and back'n'forth
conversion of the API model.

We expose the following APIs from the package for external users
to get, create/update, or delete a recorder:

  - func (r *Recorder) RetrieveRecorder(id ID) (*RecInfo, error)
  - func (r *Recorder) UpsertRecorder(recInfoNew *RecInfo) (bool, error)
  - func (r *Recorder) DeleteRecorder(id ID) (bool, error)

And one more API to get a list of all current recorders:

  - func (r *Recorder) RetrieveRecorderSet() []*RecInfo

A recorder object consists of a unique id, capture length and a
list of corresponding filters:

  type RecInfo struct {
	id      ID
	capLen  uint16
	filters []RecorderTuple
  }

A single filter is defined as:

  type RecorderTuple struct {
	srcPrefix net.IPNet
	srcPort   uint16
	dstPrefix net.IPNet
	dstPort   uint16
	proto     u8proto.U8proto
  }

The recorder id management is delegated to caller code currently, but
in future this could also be managed by the recorder package. Meaning,
Hubble will have an internal id generator plus an internal mapping of
id-to-pcap file. The id is not exposed to users, it is used as a means
to correlate incoming recorder perf RB events which the Hubble reader
will pick up and then push to the right pcap file.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>